requirements.txt
google-generativeai
Pillow
pytesseract
vosk
pyaudio